code/evaluate/requirements.txt
Pillow
pathlib
requests
